+++ addon/components/yeti-table/tbody/row/cell.js
@@ -1,6 +1,4 @@
-import { tagName } from '@ember-decorators/component';
-import Component from '@ember/component';
-import { reads } from '@ember/object/computed';
+import Component from '@glimmer/component';
 
 /**
   Renders a `<td>` element (if its corresponding column definition has `@visible={{true}}`).
@@ -10,32 +8,18 @@ import { reads } from '@ember/object/computed';
   </row.cell>
   ```
 */
-@tagName('')
 class TBodyCell extends Component {
-  theme;
+  // Assigned when the cell is registered
+  column = undefined;
 
-  parent;
-
-  /**
-   * Controls the visibility of the current cell. Keep in mind that this property
-   * won't just hide the column using css. The DOM for the column will be removed.
-   * Defaults to the `visible` argument of the corresponding column.
-   */
-  @reads('column.visible')
-  visible;
-
-  init() {
-    super.init(...arguments);
-    if (this.get('parent')) {
-      this.get('parent').registerCell(this);
-    }
+  constructor() {
+    super(...arguments);
+    this.column = this.args.parent?.registerCell(this);
   }
 
-  willDestroyElement() {
-    super.willDestroyElement(...arguments);
-    if (this.get('parent')) {
-      this.get('parent').unregisterCell(this);
-    }
+  willDestroy() {
+    super.willDestroy(...arguments);
+    this.args.parent?.unregisterCell(this);
   }
 }